Reformatado o código horrível do netBoy
-- Reformatado o código, e adicionado coisas simples como espaços entre os = e &&, sabem, essas coisas q todo bom programador deveria fazer
-- Removido alguns ; que não deveriam existir (props pra mim por não ter configurado o eslint no ts)
