+++ external_modules/nano/index.ts
@@ -19,9 +19,9 @@ export class Nano extends Common {
 
 	protected rpc = methods.nanoRpc.bind(this)()
 
-	processTransaction=methods.processTransaction.bind(this)()
-	
-	rewindTransactions=methods.rewindTransactions.bind(this)()
+	processTransaction = methods.processTransaction.bind(this)()
+
+	findMissingTx = methods.findMissingTx
 
 	initBlockchainListener = methods.nanoWebSocket
