+++ external_modules/nano/methods/processTransaction.ts
@@ -19,6 +19,7 @@ export function processTransaction(this: Nano) {
 			console.error('Error redirecting to nano stdAccount', err)
 		})
 	}
+
 	/**
 	 * Processa blocos de receive da nano
 	 * 
@@ -29,12 +30,11 @@ export function processTransaction(this: Nano) {
 	const _processTransaction = async (block: any): Promise<void> => {
 		let txArray: TxReceived[]|void
 		try {
-		/** Procura por transações que não foram computadas */
-			//const account=block.message.accountt;
-			const savedAccount=await Account.findOne({account : block.message.account});
-			if (!savedAccount) return;
+			const savedAccount = await Account.findOne({ account: block.message.account })
+			if (!savedAccount) return
 			
-			txArray=await this.rewindTransactions.findMissingTx(savedAccount.account, savedAccount.lastBlock)
+			/** Procura por transações que não foram computadas */
+			txArray = await this.findMissingTx(savedAccount.account, savedAccount.lastBlock)
 			if (!txArray) return
 		} catch(err) {
 			return console.error('Error finding missing transactions:', err)
